+++ www/src/pages-markdown/whatsnew.md
@@ -9,6 +9,7 @@ header: true
  - Cards shuffled by Gold Rush now properly cost (0) lun. (1683)
  - Knowledge Hoarder now actually has Dash. (1684)
  - Fixed name "Right of Worhip" -> "Right of Worship" (1685)
+ - The "Celestial Conduit" artifact now actually does its effect. (1686)
 
 ### 0.8.72-3.1.2 (April 7, 2020)